requirements.txt
Django==1.7
Pillow==2.6.0
django-ipware==0.1.0
psycopg2==2.5.4
requests==2.4.3
wsgiref==0.1.2
